+++ src/components/AccountInfo/AccountInfo.tsx
@@ -20,18 +20,18 @@ export type AccountInfoProps = {};
 const AccountInfo = (props: AccountInfoProps) => {
   const isMobile = useIsMobile();
   const maxLength = isMobile ? 12 : 16;
-  const { account } = useAccountStore();
+  const { account, user } = useAccountStore();
 
-  if (!account) {
+  if (!account || !user) {
     return null;
   }
 
   return (
     <Card>
       <CardHeader
-        title={<Truncate variant="email" text={account.email} maxLength={maxLength} />}
+        title={<Truncate variant="email" text={user.email} maxLength={maxLength} />}
         subheader={<Address variant="text" address={account.address} maxLength={maxLength} />}
-        avatar={<Avatar src={account.avatar} />}
+        avatar={<Avatar src={user.avatar} />}
         action={
           <Stack direction="row" spacing={1}>
             <CopyButton value={account.address} successMessage="Address copied" />