+++ src/components/Collectibles/index.ts
@@ -1,0 +1,3 @@
+export * from './CollectibleItemPage';
+export * from './CollectibleListPage';
+export * from './CollectibleListItem';