+++ src/components/Login/LoginPage.tsx
@@ -17,6 +17,7 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import { AuthApiService } from '~/api/auth-api.service';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { createNextUrl, getTokenWithFacebook, getTokenWithGoogle } from './auth.service';
+import { useEffect } from 'react';
 
 interface LogInProps {
   variant?: 'signin' | 'signup';
@@ -31,6 +32,11 @@ const validationSchema = yup
 export const LoginPage = ({ variant = 'signin' }: LogInProps) => {
   const location = useLocation();
   const navigate = useNavigate();
+
+  useEffect(() => {
+    const isSignUp = location.pathname.endsWith('signup');
+    localStorage.setItem('showProductTour', isSignUp.toString());
+  }, [location.pathname]);
 
   const {
     register,