+++ src/hooks/useNetworkStore.ts
@@ -1,0 +1,7 @@
+import { useWalletStore } from './useWalletStore';
+
+export const useNetworkStore = () => {
+  const walletStore = useWalletStore();
+
+  return walletStore.networkStore;
+};