+++ src/routes/WalletHome/WalletHome.tsx
@@ -1,19 +1,36 @@
-import { useState } from 'react';
+import { useCallback, useEffect } from 'react';
 import { observer } from 'mobx-react-lite';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useLocation, useNavigate, Outlet } from 'react-router-dom';
 import { Stack, ToggleButton, ToggleButtonGroup, useIsMobile } from '@cere-wallet/ui';
 
-import { AccountBalanceWidget, ActivityList, AssetList, OnboardingDialog } from '~/components';
+import { AccountBalanceWidget, OnboardingDialog } from '~/components';
 import { WalletProductTour } from '~/components/ProductTours';
+
+enum Tabs {
+  assets = 'assets',
+  activity = 'activity',
+  collectibles = 'collectibles',
+}
 
 const WalletHome = () => {
   const isMobile = useIsMobile();
   const location = useLocation();
   const navigate = useNavigate();
 
-  const [currentTab, setCurrentTab] = useState<'assets' | 'activity'>('assets');
   const showOnboarding = location.hash.slice(1) === 'onboarding';
   const showProductTour = location.hash.slice(1) === 'product-tour';
+
+  const getActiveFromLocation = useCallback(() => {
+    const tab = location.pathname.split('/')?.pop() || '';
+    return tab in Tabs ? Tabs[tab as keyof typeof Tabs] : Tabs.assets;
+  }, [location.pathname]);
+
+  useEffect(() => {
+    if (localStorage.getItem('showProductTour') !== 'false' && !showProductTour) {
+      navigate({ ...location, hash: 'product-tour' });
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
     <Stack spacing={4}>
@@ -26,18 +43,19 @@ const WalletHome = () => {
           fullWidth
           color="primary"
           size={isMobile ? 'small' : 'medium'}
-          value={currentTab}
-          onChange={(event, value) => value && setCurrentTab(value)}
+          value={getActiveFromLocation()}
+          onChange={(event, value) => value && navigate({ ...location, pathname: `${value}` })}
           sx={{
             maxWidth: 430,
             alignSelf: 'center',
           }}
         >
-          <ToggleButton value="assets">Assets</ToggleButton>
-          <ToggleButton value="activity">Activity</ToggleButton>
+          <ToggleButton value={Tabs.assets}>Assets</ToggleButton>
+          <ToggleButton value={Tabs.collectibles}>Collectibles</ToggleButton>
+          <ToggleButton value={Tabs.activity}>Activity</ToggleButton>
         </ToggleButtonGroup>
 
-        {currentTab === 'assets' ? <AssetList dense={isMobile} /> : <ActivityList dense={isMobile} />}
+        <Outlet />
       </Stack>
 
       <OnboardingDialog open={showOnboarding} onClose={() => navigate({ ...location, hash: '' })} />