+++ src/stores/EmbededWalletStore/EmbeddedWalletStore.ts
@@ -14,6 +14,7 @@ import { BalanceStore } from '../BalanceStore';
 import { ActivityStore } from '../ActivityStore';
 import { AppContextStore } from '../AppContextStore';
 import { AuthenticationStore } from '../AuthenticationStore';
+import { CollectiblesStore } from '../CollectiblesStore';
 
 export class EmbeddedWalletStore implements Wallet {
   readonly instanceId = randomBytes(16).toString('hex');
@@ -21,6 +22,7 @@ export class EmbeddedWalletStore implements Wallet {
   readonly approvalStore: ApprovalStore;
   readonly networkStore: NetworkStore;
   readonly assetStore: AssetStore;
+  readonly collectiblesStore: CollectiblesStore;
   readonly balanceStore: BalanceStore;
   readonly activityStore: ActivityStore;
   readonly appContextStore: AppContextStore;
@@ -43,6 +45,7 @@ export class EmbeddedWalletStore implements Wallet {
     this.networkStore = new NetworkStore(this);
     this.accountStore = new AccountStore(this);
     this.assetStore = new AssetStore(this);
+    this.collectiblesStore = new CollectiblesStore(this);
     this.balanceStore = new BalanceStore(this, this.assetStore);
     this.activityStore = new ActivityStore(this);
     this.appContextStore = new AppContextStore(this);
@@ -129,7 +132,7 @@ export class EmbeddedWalletStore implements Wallet {
       },
 
       onUserInfoRequest: async () => {
-        return toJS(this.accountStore.userInfo);
+        return toJS(this.accountStore.loginData?.userInfo);
       },
 
       onWindowClose: async ({ preopenInstanceId }) => {
@@ -160,11 +163,11 @@ export class EmbeddedWalletStore implements Wallet {
      * TODO: Refactor to prevent duplicated messages
      */
     reaction(
-      () => !!this.account,
+      () => !!this.accountStore.userInfo,
       (loggedIn) => {
         this.walletConnection?.setLoggedInStatus({
           loggedIn,
-          verifier: this.account?.verifier,
+          verifier: this.accountStore.userInfo?.verifier,
         });
       },
     );
@@ -184,32 +187,24 @@ export class EmbeddedWalletStore implements Wallet {
   private async setupRpcConnection() {
     await when(() => !!this.networkStore.network);
 
-    const privateKey = this.account?.privateKey;
-    const chainConfig = this.networkStore.network!;
-
-    const engine = await createWalletEngine({
-      privateKey,
-      chainConfig,
-
-      getAccounts: () => (this.account ? [this.account.address] : []),
+    const engine = createWalletEngine({
+      chainConfig: this.networkStore.network!,
+      getPrivateKey: () => this.accountStore.privateKey,
+      getAccounts: () => this.accountStore.accounts,
       onPersonalSign: (request) => this.approvalStore.approvePersonalSign(request),
       onSendTransaction: (request) => this.approvalStore.approveSendTransaction(request),
     });
 
-    createRpcConnection({ engine, logger: console });
-
-    if (this.account && this.account?.privateKey !== privateKey) {
-      await engine.setupProvider(this.account.privateKey);
-    }
-
     this.provider = new providers.Web3Provider(engine.provider);
 
-    /**
-     * Setup provider when account privateKey is changed
-     */
+    createRpcConnection({
+      engine,
+      logger: console,
+    });
+
     reaction(
-      () => this.account?.privateKey,
-      (privateKey) => engine.setupProvider(privateKey),
+      () => this.accountStore.accounts,
+      (accounts) => engine.updateAccounts(accounts),
     );
   }
 }