+++ src/stores/WalletStore/WalletStore.ts
@@ -1,7 +1,8 @@
 import { randomBytes } from 'crypto';
 import { providers } from 'ethers';
 import { makeAutoObservable, runInAction, when } from 'mobx';
-import { createProvider, Provider as EngineProvider } from '@cere-wallet/wallet-engine';
+import { createWalletEngine } from '@cere-wallet/wallet-engine';
+import { DEFAULT_NETWORK, getChainConfig } from '@cere-wallet/communication';
 
 import { Provider, Wallet, WalletStatus } from '../types';
 import { AccountStore } from '../AccountStore';
@@ -11,12 +12,14 @@ import { BalanceStore } from '../BalanceStore';
 import { ActivityStore } from '../ActivityStore';
 import { AppContextStore } from '../AppContextStore';
 import { AuthenticationStore } from '../AuthenticationStore';
+import { CollectiblesStore } from '../CollectiblesStore';
 
 export class WalletStore implements Wallet {
   readonly instanceId: string;
   readonly accountStore: AccountStore;
   readonly networkStore: NetworkStore;
   readonly assetStore: AssetStore;
+  readonly collectiblesStore: CollectiblesStore;
   readonly balanceStore: BalanceStore;
   readonly activityStore: ActivityStore;
   readonly authenticationStore: AuthenticationStore;
@@ -35,10 +38,15 @@ export class WalletStore implements Wallet {
     this.networkStore = new NetworkStore(this);
     this.accountStore = new AccountStore(this);
     this.assetStore = new AssetStore(this);
+    this.collectiblesStore = new CollectiblesStore(this);
     this.balanceStore = new BalanceStore(this, this.assetStore);
     this.activityStore = new ActivityStore(this);
     this.appContextStore = new AppContextStore(this);
     this.authenticationStore = new AuthenticationStore(this.accountStore, this.appContextStore);
+
+    if (this.isRoot()) {
+      this.networkStore.network = getChainConfig(DEFAULT_NETWORK);
+    }
   }
 
   isRoot() {
@@ -74,25 +82,22 @@ export class WalletStore implements Wallet {
   }
 
   async init() {
-    let provider: EngineProvider | undefined;
-
     await when(() => !!this.network);
 
     if (this.isRoot()) {
       await this.authenticationStore.rehydrate();
     } else {
-      await when(() => !!this.account);
+      await when(() => !!this.accountStore.privateKey);
     }
 
-    if (this.account && this.network) {
-      provider = await createProvider({
-        privateKey: this.account.privateKey,
-        chainConfig: this.network,
-      });
-    }
+    const engine = createWalletEngine({
+      chainConfig: this.network!,
+      getAccounts: () => this.accountStore.accounts,
+      getPrivateKey: () => this.accountStore.privateKey,
+    });
 
     runInAction(() => {
-      this.currentProvider = provider && new providers.Web3Provider(provider);
+      this.currentProvider = new providers.Web3Provider(engine.provider);
       this.initialized = true;
     });
   }