+++ CHANGELOG.md
@@ -4,11 +4,16 @@
 
 -
 
+### v1.8.0
+
+- [DAV-2808] Add topup button to widget
+
 ### v1.7.0
 
 - [DAV-3365] Fix microcopy Collectibles
 - [DAV-3350] Fix button jumping to bottom on desktop
 - [DAV-3155] Add overlay modals support + enable by default in SDK
+- [DAV-3653] Fix custom JSON RPC endpoint SDK option
 
 ### v1.6.0